app/requirements.txt
Flask==2.2.3
psycopg2-binary==2.9.5
gunicorn==20.1.0
